Rename debug_client -> dart_debug_client
