+++ test/flutter_only/debug/flutter_test.test.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 import * as vs from "vscode";
 import { DebugProtocol } from "vscode-debugprotocol";
 import { fsPath } from "../../../src/utils";
-import { DartDebugClient } from "../../debug_client";
+import { DartDebugClient } from "../../dart_debug_client";
 import { activate, defer, ext, flutterHelloWorldFolder, flutterTestBrokenFile, flutterTestMainFile, flutterTestOtherFile, getLaunchConfiguration, openFile, positionOf } from "../../helpers";
 
 describe("flutter test debugger", () => {
